Avoid using deprecated `distutils`
This silences deprecations when running on Python 3.10 or newer:

> The distutils package is deprecated and slated for removal in Python
> 3.12. Use setuptools or check PEP 632 for potential alternatives

The `LooseVersion` helper itself is very simple -- it splits on any `.`
characters, tries to coerce to integers, and then otherwise just
compares ordering. We could easily reproduce `LooseVersion` or choose to
use another dependency like `packaging`, but there's an easy option
available as well: just rely on the fact that `libdmtx` only releases
numeric releases, and that we only need to check for *older* releases.